requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
distro==1.9.0
groq==0.28.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.65
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.45
orjson==3.10.18
packaging==24.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
zstandard==0.23.0
